Create posts controller -  backend optimization
